+++ src/common/data/builder.tsx
@@ -142,27 +142,41 @@ export function buildShallowChartData(
   const isVertical = direction === 'vertical';
 
   for (const point of series) {
-    const x = normalizeValue(
-      isVertical ? point.key : point.data,
-      maxBigInteger
-    );
+    const isTuple = Array.isArray(point.data);
 
-    const y = normalizeValue(
-      isVertical ? point.data : point.key,
-      maxBigInteger
-    );
+    const props = {
+      k0: normalizeValue(
+        isVertical ? 0 : point.key,
+        maxBigInteger
+      ),
+      k1: normalizeValue(
+      point.key,
+        maxBigInteger
+      ),
+      v0: normalizeValue(
+        isTuple ? point.data[0] : isVertical ? 0 : point.data,
+        maxBigInteger
+      ),
+      v1: normalizeValue(
+        isTuple ? point.data[1] : point.data,
+        maxBigInteger
+      )
+    };
+
+    const xProp = isVertical ? 'k' : 'v';
+    const yProp = isVertical ? 'v' : 'k';
 
     result.push({
-      key: normalizeValueForFormatting(point.key),
-      value: normalizeValueForFormatting(point.data),
+      key: normalizeValueForFormatting(props.k1),
+      value: normalizeValueForFormatting(props.v1),
       meta: point.meta,
       id: point.id,
-      x,
-      x0: isVertical ? x : 0,
-      x1: x,
-      y,
-      y0: isVertical ? 0 : y,
-      y1: y
+      x: props[`${xProp}1`],
+      x0: props[`${xProp}0`],
+      x1: props[`${xProp}1`],
+      y: props[`${yProp}1`],
+      y0: props[`${yProp}0`],
+      y1: props[`${yProp}1`]
     });
   }